% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifscanR.R
\docType{package}
\name{motifscanR}
\alias{motifscanR}
\title{motifscanR: motifscan in R}
\description{
The motifscanR package is designed for analyzing many sequences and many
motifs to find which sequences contain which motifs.
}
\author{
Louis Kwok
}
