% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pwmType}
\alias{pwmType}
\title{pwmType}
\usage{
pwmType(pwm)
}
\arguments{
\item{pwm}{PWMatrix object}
}
\value{
'log','log2', or 'frequency' depending on type of pwm
}
\description{
Determines type of PWM
}
\keyword{internal}
